Add node-mapnik and build instructions
This commit gets the Makefile to the point where it can render a map
with a highlighted congressional district.

Mapnik is an extremely powerful and moderately fast map rendering tool
which is used to render the tiles displayed on OpenStreetMap.org. It can
handle way more than we will use it for.

Mapnik is not easy to compile. Conveniently, the node-mapnik package
downloads a pre-built version of Mapnik for most platforms, so that in
itself is a good enough reason to write everything here in Node although
there is a Ruby `mapnik` gem which does the same thing. Also, there is
more momentum around the Node ecosystem due to projects like Mapbox
Studio which are JS apps that presumable depend on node-mapnik.

As for how Mapnik works: it takes a Map XML specification containing
a list of <Layer> and <Style> definitions. Partial documentation is
here[1]. Also read [this bit about <Filter>s][2] which are helpful in
targeting shapes with certain attributes (i.e. a certain congressional
district). Or, if examples are your preferred way to see how something
works, [the XML specification used to render OpenStreetMap][3] is open
source and uses every feature under the sun.

[1]: https://github.com/mapnik/mapnik/wiki/XMLConfigReference
[2]: https://github.com/mapnik/mapnik/wiki/SymbologySupport
[3]: https://github.com/openstreetmap/mapnik-stylesheets/blob/master/osm.xml