Add build process which maps to OCD IDs
This creates a pattern where each "map" (a directory under config) can
specify an "ocdid_mapping.csv" file which relates each shapefile's
features' attributes to an OCD ID. The format of this file is that the
first _n_ columns correspond to the _n_ entries in the "render_each"
configuration for that map in config/maps.json.

For example, there might be a map in `config/maps.json` for San
Francisco supervisor districts. (This map doesn't currently exist, but
I imagine it will be one of the first we add.) The shapefiles provided
by the city probably have features that have an attribute "district"
with the value being each district's number.

The ocdid_mapping.csv file would look something like this (assuming
there are 12 districts):

1,ocd-division/state:ca/place:san_francisco/council_district:1
2,ocd-division/state:ca/place:san_francisco/council_district:2
3,ocd-division/state:ca/place:san_francisco/council_district:3
4,ocd-division/state:ca/place:san_francisco/council_district:4
[...and so on, through 12...]

If this file is dynamically generated, add some script into the map's
config directory (like config/114_congress/generateOcdidMapping.js) and
add its invocation to the top-level Makefile.