Clean up congress maps on US borders
The "countries" shapefile map of the USA is at a lower resolution than
the other sources, which was leading to some rough edges where
boundaries didn't match up.

I've hidden the country rendering for the US, which cleans up these
edges. To fill the rest of the country back in, I've added a fill to the
"usa_states" layer.