Constrain SLDL/SLDU maps to land
The SLDL/SLDU district shapefiles aren't constrained to land, which
means that we lose much of the recognizability of districts along
coastlines. (Notably, San Francisco's districts look particularly
janky).

If we remove the map background from the Mapnik render, we can use the
state shapefile as a mask and only render the parts of the districts
that are on land. I spent a while trying to figure out how to use mapnik
compositing to add the background back in afterwards, but after some
time decided to move to something simpler: using a simple GraphicsMagick
transform to replace the transparent background with the expected water
color.

This isn't really ideal, as it adds a dependency and nearly doubles the
time for the render, but right now we're more constrained by developer
time than render time, so I'm going to #ShipIt