Record cloudinary upload results
We're uploading a lot of images now, and creating new versions of images
that I have already uploaded. The uploader will now write information
about each file uploaded to stdout, and about upload failures to stderr.

I also bumped the number of concurrent uploads, as this was getting
slow.

Recommended usage:
  node bin/upload.js | pv -ls 28180 > uploads.csv

(where 28180 is the number of images I expect to upload)