Add roundCoordinates function/module
This function takes an arbitrarily nested array of arrays of floating
point numbers (coordinates) and returns a version of those numbers that
are rounded within a certain tolerance (EPSILON).

This will be useful for coordinates that look like "-47.123450000001".
Although this coordinate probably resulted from a floating-point
representation limitation, when serialized in GeoJSON it will be
represented as a series of UTF-8 bytes (each digit), and thus we are not
bound to the same representation limitations as floating point numbers
are.