Replace "simplify-geojson" with "mapshaper" simplification
The simplify-geojson module was outputting simplified polygons
containing only a single point. This is invalid GeoJSON and breaks the
iOS GeoJSON parser.

Conveniently, other options exist in this space - there is a library
called mapshaper which is quite featureful and perhaps we will want to
use it in other ways. Its simplification algorithm is more configurable
and also supports truncating the precision of floating-point lat/long
values out-of-the-box (goodbye, lib/roundCoordinates.js!)

However, I had considerable difficulty using the javascript API of
mapshaper. I could not figure out the invocation of
mapshaper.runCommands() or mapshaper.applyCommands() to successfully
simplify the map in-process. I spent a few hours trying, to no avail, so
this commit just shells out to the mapshaper executable, which somehow
works much better. Its slowness should be low hanging fruit to improve
subsequently.

As for the simplification options, I opted us into the `visvalingam`
simplification algorithm, which looks better on the simplification page
here[1]. The 10% parameter and the 0.0001 precision arguments were
determined somewhat arbitrarily based on looking at output maps.

[1]: https://github.com/mbloch/mapshaper/wiki/Simplification-Tips