Fix district map counties including water
The county shapefiles we use to generate the GeoJSON include water.

This was a problem we initially solved at render-time through clever use
of the "src-atop" rendering option. This worked pretty well at the time,
but now that we need to also output GeoJSON shapes for these counties
with the same geometry, we'll need to do this earlier in the rendering
pipeline.

This adds a couple "computed" shapefiles that are output by running
`mapshaper` commands to clip / filter existing shapes.

[Fixes API-1496] - Fix district map counties including water