Don't skip processing counties that no longer have governments
Initially, we had skipped features in the county shapefile with
"FUNCSTAT" == "N" because it seemed like the description
("Nonfunctioning legal entity") would imply that the feature is not
really a thing.

It turns out that that label is applied to counties which have abolished
the county government in favor of letting cities do their own thing. For
instance, [Essex County, Massachusetts][1].

However, it turns out that these counties are still quite useful to us
because we render the state map as the union of all the county shapes.
So even though these non-counties will never have elected officials, we
still need to render them.

[Fixes WEB-1912] - Massachusetts County Info Missing for Presidential
                   Race
[1]: https://en.wikipedia.org/wiki/Essex_County,_Massachusetts