Include "Fictitious entities" in OCDID processing
The Census creates "fictitious entities" for purposes of producing a
shapefile that has no area missing.

This is most notably the case in the Commonwealth of Virginia -- so
called a "Commonwealth" because it identifies itself more as a
collection of cities than a cohesive state. This identity is represented
by the unusually large number of independent cities (cities that are not
part of a county).

For each of these cities, there is an entry with "FUNCSTAT" = "F".

It's likely we will need some more processing of these entities. There
is currently an OCDID alias to each corresponding place (city). For
example, "county:alexandria" has an alias to "place:alexandria", so we
will probably handle these areas in the same way as we handle
consolidated city-counties.